requirements.txt
streamlit
pandas
altair
vl-convert-python
sqlalchemy
psycopg2-binary
python-dotenv